+++ src/components/book/Cover.styl
@@ -1,0 +1,11 @@
+@import '../vars.styl';
+
+article.cover
+  align-items center
+  display flex
+  height coverHeight
+  justify-content center
+  width coverWidth
+
+  img
+    width coverImageWidth
