+++ src/components/book/DetailedCover.styl
@@ -1,14 +1,4 @@
-@import './vars.styl';
-
-article.cover
-  align-items center
-  display flex
-  height coverHeight
-  justify-content center
-  width coverWidth
-
-  img
-    width coverImageWidth
+@import '../vars.styl';
 
 article.detailedCover
   height detailedCoverHeight
