+++ src/components/common/Button.styl
@@ -1,4 +1,4 @@
-@import './vars.styl';
+@import '../vars.styl';
 
 buttonHeight = 60px
