+++ src/components/common/Loader.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PureRenderMixin from 'react-addons-pure-render-mixin';
 
-import '../../assets/styl/loader.styl';
+import './Loader.styl';
 
 export default React.createClass({
   displayName: 'Loader',
