+++ src/components/common/Progressbar.styl
@@ -1,4 +1,4 @@
-@import './vars.styl';
+@import '../vars.styl';
 
 .progressbar
   align-items center
