+++ src/components/common/Sidebar.styl
@@ -1,4 +1,4 @@
-@import './vars.styl';
+@import '../vars.styl';
 
 aside.sidebar
   align-items flex-end
