+++ src/components/drawer/Backdrop.styl
@@ -1,4 +1,4 @@
-@import './vars.styl';
+@import '../vars.styl';
 
 .backdrop
   background rgba(0,0,0,0.6)
