+++ src/components/library/Library.styl
@@ -1,4 +1,4 @@
-@import './vars.styl';
+@import '../vars.styl';
 
 main.library
   display flex
