+++ src/components/library/Shelf.styl
@@ -1,4 +1,4 @@
-@import './vars.styl';
+@import '../vars.styl';
 
 book-space = 15px
