+++ src/components/vars.styl
@@ -19,6 +19,10 @@ coverImageWidth = 140px
 dimOpacity = 0.4
 dimFontColour = alpha(white, dimOpacity)
 
+drawerBuffer = 200px
+drawerWidth = 570px
+drawerHeaderHeight = 250px
+
 angled-content(color, height)
   content ''
   display block
